requirements.txt
requests
pandas
python-dateutil
openai